Simplifying the interface (work in progress)
I came back to this library after one year and there are
lots of things to do.
First thing I noticed is that this is akin to a
gone wrong, reduced copy of the ghc TcM monad.
It has some nice properties because it uses the ST
monad, but other than that, it is a kludge.

My goal is to bring it a bit on par with the TcM
monad in simplicity of use, copying the distintion
between Tyvars and TcTyvars in there for instance.

Another very interesting option is to try to implement
the unification algorithm seen in the Handbook in a
purely functional way. This would be really awesome if
it is fast enough

darcs-hash:20070525094446-bec41-a7d42161deb4e035cbdec422c214a7e6cbb0bc80.gz